+++ locale_emoji/lib/src/locale_emoji.dart
@@ -13,31 +13,42 @@ import 'utils.dart';
 /// 127397 == 0x1F1A
 const _offset = 0x1F1E6 - 0x41;
 
+/// Pattern for country code according to ISO 3166-1 (alpha-2 code)
+const _countryPattern = r"^[A-Z]{2}$";
+
 /// Get Flag Emoji from the locale information.
 ///
 /// Casing of input parameters is can by any (lower, UPPER, mIxEd).
-String? getFlagEmoji(
-  String languageCode, {
+String? getFlagEmoji({
+  String? languageCode,
   String? scriptCode,
   String? countryCode,
 }) {
-  final String language;
-  if (scriptCode != null && scriptCode.isNotEmpty) {
-    language = '${languageCode.toLowerCase()}-${scriptCode.capitalized()}';
-  } else {
-    language = languageCode.toLowerCase();
-  }
-
   String? country = countryCode?.toUpperCase();
 
-  if (country == null || !RegExp(r"^[A-Z]{2}$").hasMatch(country)) {
-    country = languageCodeToCountryCode[language];
+  if (country == null || !RegExp(_countryPattern).hasMatch(country)) {
+    country = _countryFromLanguage(languageCode, scriptCode);
   }
 
   if (country == null) {
     return null;
   }
 
-  return String.fromCharCode(country.codeUnitAt(0) + _offset) +
-      String.fromCharCode(country.codeUnitAt(1) + _offset);
+  return _countryToFlagEmoji(country);
 }
+
+/// Determines whether country code can be obtained from language code
+String? _countryFromLanguage(String? languageCode, String? scriptCode) {
+  if (languageCode == null) return null;
+
+  final language = scriptCode != null && scriptCode.isNotEmpty
+      ? '${languageCode.toLowerCase()}-${scriptCode.capitalized()}'
+      : languageCode.toLowerCase();
+
+  return languageCodeToCountryCode[language];
+}
+
+/// Converts country code to the actual Flag Emoji
+String _countryToFlagEmoji(String country) =>
+    String.fromCharCode(country.codeUnitAt(0) + _offset) +
+    String.fromCharCode(country.codeUnitAt(1) + _offset);